add some tests to timeout code
