--- src/platform/web/dom/request/timeout.js
@@ -1,51 +1,0 @@
-/*
-Copyright 2020 Bruno Windels <bruno@windels.cloud>
-Copyright 2020 The Matrix.org Foundation C.I.C.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-import {
-    AbortError,
-    ConnectionError
-} from "../../../../matrix/error.js";
-
-
-export function abortOnTimeout(createTimeout, timeoutAmount, requestResult, responsePromise) {
-    const timeout = createTimeout(timeoutAmount);
-    // abort request if timeout finishes first
-    let timedOut = false;
-    timeout.elapsed().then(
-        () => {
-            timedOut = true;
-            requestResult.abort();
-        },
-        () => {}    // ignore AbortError when timeout is aborted
-    );
-    // abort timeout if request finishes first
-    return responsePromise.then(
-        response => {
-            timeout.abort();
-            return response;
-        },
-        err => {
-            timeout.abort();
-            // map error to TimeoutError
-            if (err.name === "AbortError" && timedOut) {
-                throw new ConnectionError(`Request timed out after ${timeoutAmount}ms`, true);
-            } else {
-                throw err;
-            }
-        }
-    );
-}
